+++ Runtime/SimpleListView.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 using Unity.Properties;
 using UnityEngine;
 using UnityEngine.Scripting.APIUpdating;
@@ -11,7 +12,47 @@ namespace Strayfarer.UI {
     [MovedFrom(true, "Retropair.UXML", "Retropair")]
     public sealed partial class SimpleListView : BindableElement {
 
+        public SimpleListView() {
+            AddToClassList("simpleList-root");
+        }
+
+        Func<VisualElement> _instantiateItem;
+        public Func<VisualElement> instantiateItem {
+            get => _instantiateItem;
+            set {
+                if (_instantiateItem != value) {
+                    _instantiateItem = value;
+                    Rebuild(true);
+                }
+            }
+        }
         public event Action<VisualElement> onInstantiateItem;
+        public event Action<VisualElement, object> onBindItem;
+
+        readonly Stack<VisualElement> pool = new();
+
+        bool InstantiateItem(out VisualElement item) {
+            if (pool.TryPop(out item)) {
+                return false;
+            }
+
+            if (_instantiateItem is not null) {
+                item = _instantiateItem();
+                return true;
+            }
+
+            if (_itemTemplate) {
+                item = _itemTemplate.Instantiate();
+                return true;
+            }
+
+            item = new VisualElement();
+            return true;
+        }
+
+        void ReturnItem(VisualElement element) {
+            pool.Push(element);
+        }
 
         VisualTreeAsset _itemTemplate;
         [UxmlAttribute]
@@ -19,7 +60,7 @@ namespace Strayfarer.UI {
             get => _itemTemplate;
             set {
                 _itemTemplate = value;
-                Rebuild();
+                Rebuild(true);
             }
         }
 
@@ -28,12 +69,8 @@ namespace Strayfarer.UI {
         public IList itemsSource {
             get => _itemsSource;
             set {
-                if (_itemsSource == value) {
-                    Update();
-                } else {
-                    _itemsSource = value;
-                    Rebuild();
-                }
+                _itemsSource = value;
+                Rebuild(false);
             }
         }
 
@@ -48,28 +85,22 @@ namespace Strayfarer.UI {
                 }
 
                 _elementsPerSection = value;
-                Rebuild();
+                Rebuild(false);
             }
         }
 
-        void Update() {
-            int newSize = _itemsSource is null
-                ? -1
-                : _itemsSource.Count;
-
-            if (builtSize != newSize) {
-                Rebuild();
-            }
-        }
+        public IEnumerable<VisualElement> items => _elementsPerSection == 0
+            ? Children()
+            : sections.SelectMany(section => section.Children());
 
         int builtSize = -1;
         readonly List<VisualElement> sections = new();
-        VisualElement GetSectionForElement(int elementIndex) {
-            if (elementsPerSection == 0) {
+        VisualElement GetSectionForItem(int itemIndex) {
+            if (_elementsPerSection == 0) {
                 return this;
             }
 
-            int sectionIndex = elementIndex / elementsPerSection;
+            int sectionIndex = itemIndex / _elementsPerSection;
             for (int i = sections.Count; i < sectionIndex + 1; i++) {
                 var section = new VisualElement();
                 section.AddToClassList($"simpleList-section");
@@ -81,52 +112,73 @@ namespace Strayfarer.UI {
             return sections[sectionIndex];
         }
 
-        VisualElement InstantiateItem() {
-            var item = _itemTemplate is null
-                ? new VisualElement()
-                : _itemTemplate.Instantiate();
-            item.AddToClassList($"simpleList-item");
-            return item;
-        }
-
-        void Rebuild() {
-            sections.Clear();
-
-            Clear();
+        public void Rebuild(bool discardItems = false) {
+            if (discardItems) {
+                pool.Clear();
+                sections.Clear();
+                Clear();
+            }
 
             if (_itemsSource is null) {
                 builtSize = -1;
+                foreach (var child in items) {
+                    ReturnItem(child);
+                }
+
+                foreach (var child in pool) {
+                    child.RemoveFromHierarchy();
+                }
+
                 return;
+            }
+
+            var previousItems = new List<VisualElement>(items);
+
+            foreach (var previousItem in previousItems.Skip(_itemsSource.Count)) {
+                ReturnItem(previousItem);
+                previousItem.RemoveFromHierarchy();
             }
 
             builtSize = 0;
 
             for (int i = 0; i < _itemsSource.Count; i++) {
-                var element = InstantiateItem();
+                if (previousItems.ElementAtOrDefault(i) is not VisualElement element) {
+                    if (InstantiateItem(out element)) {
+                        element.AddToClassList($"simpleList-item");
+
+                        try {
+                            onInstantiateItem?.Invoke(element);
+                        } catch (Exception e) {
+                            Debug.LogException(e);
+                        }
+                    }
+                }
 
                 if (i == 0) {
                     element.AddToClassList("first-child");
+                } else {
+                    element.RemoveFromClassList("first-child");
                 }
 
                 if (i == _itemsSource.Count - 1) {
                     element.AddToClassList("last-child");
+                } else {
+                    element.RemoveFromClassList("last-child");
                 }
 
-                element.dataSource = _itemsSource[i];
-                GetSectionForElement(i).Add(element);
+                GetSectionForItem(i).Add(element);
+
+                if (element.dataSource != _itemsSource[i]) {
+                    element.dataSource = _itemsSource[i];
+                    try {
+                        onBindItem?.Invoke(element, _itemsSource[i]);
+                    } catch (Exception e) {
+                        Debug.LogException(e);
+                    }
+                }
 
                 builtSize++;
-
-                try {
-                    onInstantiateItem?.Invoke(element);
-                } catch (Exception e) {
-                    Debug.LogException(e);
-                }
             }
-        }
-
-        public SimpleListView() {
-            AddToClassList("simpleList-root");
         }
     }
 }